+++ source/CHOverlayView.h
@@ -6,6 +6,7 @@
 //  Copyright 2010 zachwaugh.com. All rights reserved.
 //
 
+#import <Cocoa/Cocoa.h>
 
 #define TAB_KEY 48
 #define Q_KEY 12
@@ -23,7 +24,6 @@ typedef enum
 	CHResizeLeftCenter
 } CHResizeDirection;
 
-#import <Cocoa/Cocoa.h>
 
 @interface CHOverlayView : NSView
 {
@@ -34,35 +34,43 @@ typedef enum
 	NSMutableDictionary *textAttrs;
 	NSMutableDictionary *smallTextAttrs;
 	
-	NSColor *fillColor;
+  NSColor *fillColor;
+	NSColor *primaryColor;
+  NSColor *alternateColor;
 	
 	CHResizeDirection resizeDirection;
 	
 	float fillOpacity;
 	
-	BOOL alternateColor;
+	BOOL switchedColors;
 	BOOL dragging;
 	BOOL drawing;
+  BOOL hovering;
 	BOOL resizing;
 	BOOL shiftPressed;
 	BOOL commandPressed;
 	BOOL lastPointInOverlay;
+  BOOL showDimensionsOutside;
 }
 
 @property(assign) NSPoint startPoint;
 @property(assign) NSPoint lastPoint;
 @property(assign) NSRect overlayRect;
 @property(retain) NSColor *fillColor;
+@property(retain) NSColor *primaryColor;
+@property(retain) NSColor *alternateColor;
 @property(retain) NSMutableDictionary *textAttrs;
 @property(retain) NSMutableDictionary *smallTextAttrs;
 @property(assign) CHResizeDirection resizeDirection;
 @property(assign) float fillOpacity;
 @property(assign, getter=isDragging) BOOL dragging;
+@property(assign, getter=isHovering) BOOL hovering;
 @property(assign, getter=isDrawing) BOOL drawing;
 @property(assign, getter=isResizing) BOOL resizing;
 @property(assign, getter=isShiftPressed) BOOL shiftPressed;
 @property(assign, getter=isCommandPressed) BOOL commandPressed;
-@property(assign, getter=isAlternateColor) BOOL alternateColor;
+@property(assign) BOOL switchedColors;
 @property(assign) BOOL lastPointInOverlay;
+@property(assign) BOOL showDimensionsOutside;
 
 @end