+++ source/CHPreferences.h
@@ -10,10 +10,19 @@
 
 
 @interface CHPreferences : NSObject
-{
-
-}
 
 + (void)registerDefaults;
++ (NSString *)copyFormat;
++ (NSColor *)primaryOverlayColor;
++ (NSColor *)alternateOverlayColor;
++ (void)setLastColor:(NSColor *)color;
++ (NSColor *)lastOverlayColor;
++ (BOOL)switchedColors;
++ (void)setSwitchedColors:(BOOL)switched;
++ (BOOL)showDimensionsOutside;
++ (NSInteger)globalHotKeyCode;
++ (NSInteger)globalHotKeyFlags;
++ (void)setGlobalHotKeyCode:(NSInteger)keyCode;
++ (void)setGlobalHotKeyFlags:(NSInteger)flags;
 
 @end