+++ source/CHPreferencesController.h
@@ -7,10 +7,20 @@
 //
 
 #import <Cocoa/Cocoa.h>
+#import <ShortcutRecorder/ShortcutRecorder.h>
 
-
-@interface CHPreferencesController : NSWindowController {
-
+@interface CHPreferencesController : NSWindowController
+{
+  IBOutlet NSColorWell *primaryColorWell;
+  IBOutlet NSColorWell *alternateColorWell;
+  IBOutlet SRRecorderControl *shortcutRecorder;
 }
+
+@property (retain) NSColorWell *primaryColorWell;
+@property (retain) NSColorWell *alternateColorWell;
+@property (retain) SRRecorderControl *shortcutRecorder;
+
+- (void)colorUpdated:(id)sender;
+- (void)toggleStartAtLogin:(id)sender;
 
 @end