+++ resources/CHANGELOG.md
@@ -1,10 +1,18 @@
-0.9.5b
+1.0b
 - Delete key now clears overlay
 - Spacebar takes screenshot
 - Fixed text shadow color on white overlay
 - Fixed issue where arrow cursor would show instead of crosshairs
-- Hides handles on resize
+- Hides handles on resize and when not hovering over overlay
 - Fixed some resizing bugs
+- Updated crosshairs cursor
+- Moved dimensions outside of overlay when too small to fit inside
+- Added user preferences
+  - Can now change overlay colors
+  - Can specify format format for copying dimensions to clipboard
+  - Can choose to always display dimensions outside overlay instead of just when it's too small
+  - Can choose global hotkey to use instead of cmd-shift-2
+  - Option to automatically start app at login
 
 0.9.1b
 - Changed "Activate" to "Show" in dropdown menu